refactor(audio-io): aec.rs をモジュール分割 (mixer / nlms)
1242行あった aec.rs を関心事ごとに分割(挙動・公開APIは非変更):
- src/aec/mixer.rs   : ReferenceMixer(far-end ミキサ)+ そのテスト
- src/aec/nlms.rs    : DelayEstimator + Aec(NLMS+残差サプレッサ)+ 各定数
                       + EchoCanceller 実装 + アルゴリズムのテスト
- src/aec.rs(ルート): モジュール宣言/再エクスポート + EchoCanceller
                       トレイト + CancellerStats + reference_mixer_task /
                       aec_task(配線)
コードは移動のみ。テスト 22+4 通過、clippy クリーン、speex ビルドも通過。
